backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.0
